# openapi/integration.json
{
  "components": {
    "schemas": {
...
      "CategoryDataRecordsPublic": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/CategoryDataRecordPublic"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "count"
        ],
        "title": "CategoryDataRecordsPublic",
        "type": "object"
      },
      "CategoryDataSchemaPublic": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          },
          "schema": {
            "additionalProperties": true,
            "type": "object"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "version",
          "schema",
          "is_active",
          "created_at"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "IntegrationApiKey": {
        "description": "An integration connection key, for example fdg_live_\u2026",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Integration API for Oblidog Ledger.",
    "title": "Oblidog Integration API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/integration/category/data-records": {
      "get": {
        "operationId": "integration-read_integration_category_data_records",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDataRecordsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Integration Category Data Records",
        "tags": [
          "integration"
        ]
      },
      "post": {
        "operationId": "integration-create_integration_category_data_record",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationCategoryDataRecordCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDataRecordPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Create Integration Category Data Record",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/category/data-records/latest": {
      "get": {
        "operationId": "integration-read_latest_integration_category_data_record",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDataRecordPublic"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Latest Integration Category Data Record",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/category/schema": {
      "get": {
        "operationId": "integration-read_integration_category_data_schema",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDataSchemaPublic"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Integration Category Data Schema",
        "tags": [
          "integration"
        ]
      }
    },
...
    "/api/v1/integration/obligations": {
      "get": {
        "operationId": "integration-read_integration_obligations",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 9999,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 12,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "lifecycle",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ObligationLifecycle"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lifecycle"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Integration Obligations",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}": {
      "get": {
        "operationId": "integration-read_integration_obligation",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Integration Obligation",
        "tags": [
          "integration"
        ]
      },
      "patch": {
        "operationId": "integration-update_integration_obligation",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObligationIntegrationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Update Integration Obligation",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}/cancel": {
      "post": {
        "operationId": "integration-cancel_integration_obligation",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Cancel Integration Obligation",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}/components": {
      "get": {
        "operationId": "integration-read_integration_obligation_components",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationComponentsPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Read Integration Obligation Components",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}/components/upsert": {
      "put": {
        "operationId": "integration-upsert_integration_obligation_component",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationObligationComponentUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationComponentUpsertResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Upsert Integration Obligation Component",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}/error": {
      "post": {
        "operationId": "integration-mark_integration_obligation_error",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Mark Integration Obligation Error",
        "tags": [
          "integration"
        ]
      }
    },
    "/api/v1/integration/obligations/{period}/mark-paid": {
      "post": {
        "operationId": "integration-mark_integration_obligation_paid",
        "parameters": [
          {
            "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
            "in": "path",
            "name": "period",
            "required": true,
            "schema": {
              "description": "Billing period in YYYY-MM format. Full obligation keys are temporarily accepted for client migration.",
              "examples": [
                "2026-09"
              ],
              "title": "Period",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligationPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "IntegrationApiKey": []
          }
        ],
        "summary": "Mark Integration Obligation Paid",
        "tags": [
          "integration"
        ]
      }
    },
...
  }
}